+++ extensions/notion/src/components/databaseViews/DatabaseListView.tsx
@@ -1,10 +1,18 @@
 import { List } from "@raycast/api";
-import { PageListItem } from "..";
+import { PageListItem } from "../PageListItem";
 import { DatabaseViewProps } from "./types";
 
 export function DatabaseListView(props: DatabaseViewProps): JSX.Element {
   // Get database page list info
-  const { databaseId, databasePages, databaseProperties, databaseView, onForceRerender, saveDatabaseView } = props;
+  const {
+    databaseId,
+    databasePages,
+    databaseProperties,
+    databaseView,
+    onPageUpdated,
+    onPageCreated,
+    saveDatabaseView,
+  } = props;
 
   return (
     <List.Section key="database-view-list" title="Recent">
@@ -16,7 +24,8 @@ export function DatabaseListView(props: DatabaseViewProps): JSX.Element {
             databaseView={databaseView}
             databaseProperties={databaseProperties}
             saveDatabaseView={saveDatabaseView}
-            onForceRerender={onForceRerender}
+            onPageUpdated={onPageUpdated}
+            onPageCreated={onPageCreated}
           />
         );
       })}