+++ extensions/notion/src/components/forms/CreateDatabaseForm.tsx
@@ -1,38 +1,44 @@
-import { ActionPanel, Icon, Detail, Form, showToast, useNavigation, Action, Image, Toast, open } from "@raycast/api";
+import { ActionPanel, Icon, Detail, Form, showToast, useNavigation, Action, Image, Toast } from "@raycast/api";
 import { useEffect, useState } from "react";
+import { useAtom } from "jotai";
 import {
-  Database,
-  DatabaseProperty,
-  DatabasePropertyOption,
-  DatabaseView,
-  User,
-  Page,
   fetchDatabases,
   fetchDatabaseProperties,
   queryDatabase,
   createDatabasePage,
-  fetchExtensionReadMe,
   notionColorToTintColor,
   fetchUsers,
-} from "../utils/notion";
-import { ActionSetVisibleProperties } from "./";
+} from "../../utils/notion";
+import { DatabaseProperty, DatabasePropertyOption, DatabaseView, Page } from "../../utils/types";
+import { ActionSetVisibleProperties } from "../actions";
+import { handleOnOpenPage } from "../../utils/openPage";
 import {
-  storeDatabases,
-  loadDatabases,
-  storeDatabaseProperties,
-  loadDatabaseProperties,
-  storeDatabasePages,
-  loadDatabasePages,
-  storeUsers,
-  loadUsers,
-  storeDatabaseView,
-  loadDatabaseView,
-} from "../utils/local-storage";
-import { useForceRenderer } from "../utils/useForceRenderer";
-import { handleOnOpenPage } from "../utils/openPage";
+  usersAtom,
+  databasesAtom,
+  databaseViewsAtom,
+  databasePropertiesAtom,
+  recentlyOpenedPagesAtom,
+} from "../../utils/state";
 
-export function CreateDatabaseForm(props: { databaseId?: string; onForceRerender?: () => void }): JSX.Element {
+export function CreateDatabaseForm(props: { databaseId?: string; onPageCreated?: (page: Page) => void }): JSX.Element {
   const presetDatabaseId = props.databaseId;
+
+  const [databaseId, setDatabaseId] = useState<string | null>(presetDatabaseId ? presetDatabaseId : null);
+  const [databaseView, setDatabaseView] = useAtom(databaseViewsAtom(databaseId || "__no_id__"));
+  const [databaseProperties, setDatabaseProperties] = useAtom(databasePropertiesAtom(databaseId || "__no_id__"));
+  const [users, storeUsers] = useAtom(usersAtom);
+  const [databases, storeDatabases] = useAtom(databasesAtom);
+  const [recentlyOpenedPages, storeRecentlyOpenedPage] = useAtom(recentlyOpenedPagesAtom);
+  const [relationsPages, setRelationsPages] = useState<Record<string, Page[]>>(
+    databaseProperties
+      .filter((cdp) => cdp.type === "relation" && cdp.relation_id)
+      .reduce((prev, cdp) => {
+        prev[cdp.id] = recentlyOpenedPages.filter((x) => x.parent_database_id === cdp.relation_id);
+        return prev;
+      }, {} as Record<string, Page[]>)
+  );
+  const [isLoadingDatabases, setIsLoadingDatabases] = useState(true);
+  const [isLoading, setIsLoading] = useState(true);
 
   // On form submit function
   const { pop } = useNavigation();
@@ -46,74 +52,31 @@ export function CreateDatabaseForm(props: { databaseId?: string; onForceRerender
     setIsLoading(true);
     const page = await createDatabasePage(values);
     setIsLoading(false);
-    if (!page) {
-      showToast({
-        style: Toast.Style.Failure,
-        title: "Couldn't create database page",
-      });
-    } else {
-      showToast({
-        title: "Page created!",
-      });
-      props.onForceRerender?.();
+    if (page) {
+      props.onPageCreated?.(page);
 
       if (openPage) {
-        handleOnOpenPage(page);
+        handleOnOpenPage(page, storeRecentlyOpenedPage);
       } else {
         pop();
       }
     }
   }
 
-  // Setup useState objects
-  const [databases, setDatabases] = useState<Database[]>([]);
-  const [databaseProperties, setDatabaseProperties] = useState<DatabaseProperty[]>([]);
-  const [databaseId, setDatabaseId] = useState<string | null>(presetDatabaseId ? presetDatabaseId : null);
-  const [databaseView, setDatabaseView] = useState<DatabaseView>();
-  const [didRerender, forceRerender] = useForceRenderer();
-  const [markdown, setMarkdown] = useState<string>();
-  const [relationsPages, setRelationsPages] = useState<Record<string, Page[]>>({});
-  const [users, setUsers] = useState<User[]>();
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-
-  // Currently supported properties
-  const supportedPropTypes = [
-    "title",
-    "rich_text",
-    "number",
-    "url",
-    "email",
-    "phone_number",
-    "date",
-    "checkbox",
-    "select",
-    "multi_select",
-    "relation",
-    "people",
-  ];
-
   // Fetch databases
   useEffect(() => {
     const fetchData = async () => {
       if (presetDatabaseId) {
-        setDatabaseId(presetDatabaseId);
+        setIsLoadingDatabases(false);
         return;
-      }
-
-      const cachedDatabases = await loadDatabases();
-
-      if (cachedDatabases) {
-        setDatabases(cachedDatabases);
       }
 
       const fetchedDatabases = await fetchDatabases();
 
-      setIsLoading(false);
-
-      if (fetchedDatabases) {
-        setDatabases(fetchedDatabases);
-        storeDatabases(fetchedDatabases);
+      if (fetchedDatabases.length) {
+        await storeDatabases(fetchedDatabases);
       }
+      setIsLoadingDatabases(false);
     };
     fetchData();
   }, []);
@@ -124,97 +87,44 @@ export function CreateDatabaseForm(props: { databaseId?: string; onForceRerender
       if (databaseId) {
         setIsLoading(true);
 
-        // load from the cache first
-
-        const cachedDatabaseProperties = await loadDatabaseProperties(databaseId);
-        setDatabaseProperties(cachedDatabaseProperties);
-
-        cachedDatabaseProperties.forEach(async function (cdp: DatabaseProperty) {
-          if (cdp.type === "relation" && cdp.relation_id && !relationsPages[cdp.relation_id]) {
-            const cachedRelationPages = await loadDatabasePages(databaseId);
-            setRelationsPages((relationsPages) => ({
-              ...relationsPages,
-              [cdp.relation_id!]: cachedRelationPages,
-            }));
-          }
-        });
-
-        let hasPeopleProperty = cachedDatabaseProperties.some((cdp) => cdp.type === "people");
-        if (hasPeopleProperty) {
-          const cachedUsers = await loadUsers();
-          setUsers(cachedUsers);
-        }
-
-        // then we fetch
-
         const fetchedDatabaseProperties = await fetchDatabaseProperties(databaseId);
-        const supportedDatabaseProperties = fetchedDatabaseProperties.filter(function (property) {
-          return supportedPropTypes.includes(property.type);
-        });
-        setDatabaseProperties(supportedDatabaseProperties);
-
-        // Fetch relation pages
-        supportedDatabaseProperties.forEach(async function (cdp: DatabaseProperty) {
-          if (cdp.type === "relation" && cdp.relation_id) {
-            const fetchedRelationPages = await queryDatabase(cdp.relation_id, undefined);
-            if (fetchedRelationPages.length > 0) {
-              setRelationsPages((relationsPages) => ({ ...relationsPages, [cdp.relation_id!]: fetchedRelationPages }));
-              storeDatabasePages(cdp.relation_id, fetchedRelationPages);
-            }
-          }
-        });
-
-        // Fetch users
-        hasPeopleProperty = supportedDatabaseProperties.some((cdp) => cdp.type === "people");
-        if (hasPeopleProperty) {
-          const fetchedUsers = await fetchUsers();
-          setUsers(fetchedUsers);
-          await storeUsers(fetchedUsers);
+        if (fetchedDatabaseProperties.length) {
+          await setDatabaseProperties(fetchedDatabaseProperties);
         }
-
-        await storeDatabaseProperties(databaseId, supportedDatabaseProperties);
 
         setIsLoading(false);
       }
     };
     fetchData();
-  }, [databaseId, didRerender]);
+  }, [databaseId, databaseView]);
 
-  // Load database view
   useEffect(() => {
-    const getDatabaseView = async () => {
-      if (!databaseId) {
-        return;
-      }
+    if (!databaseId) {
+      return;
+    }
 
-      const loadedDatabaseView = await loadDatabaseView(databaseId);
+    setIsLoading(true);
 
-      if (loadedDatabaseView?.create_properties) {
-        setDatabaseView(loadedDatabaseView);
-      } else {
-        const dbView = { ...(loadedDatabaseView || {}), create_properties: [] as string[] };
-        databaseProperties?.forEach(function (dp) {
-          dbView.create_properties.push(dp.id);
-        });
-        setDatabaseView(dbView);
+    // Fetch relation pages
+    databaseProperties.forEach(async function (cdp: DatabaseProperty) {
+      if (cdp.type === "relation" && cdp.relation_id) {
+        const fetchedRelationPages = await queryDatabase(cdp.relation_id, undefined);
+        if (fetchedRelationPages.length > 0) {
+          setRelationsPages((relationsPages) => ({ ...relationsPages, [cdp.relation_id!]: fetchedRelationPages }));
+        }
       }
-    };
-    getDatabaseView();
+    });
+
+    // Fetch users
+    if (databaseProperties.some((cdp) => cdp.type === "people")) {
+      fetchUsers().then(storeUsers);
+    }
+
+    setIsLoading(false);
   }, [databaseProperties]);
 
-  // Fetch Notion Extension README
-  useEffect(() => {
-    const fetchREADME = async () => {
-      if (!markdown) {
-        const fetchedREADME = await fetchExtensionReadMe();
-        setMarkdown(fetchedREADME);
-      }
-    };
-    fetchREADME();
-  }, []);
-
-  if (databases && !databases[0]) {
-    return <Detail markdown={markdown} isLoading={!markdown} />;
+  if (!isLoading && !isLoadingDatabases && !databases.length) {
+    return <Detail markdown={`No databases`} />;
   }
 
   // Handle save new database view
@@ -222,11 +132,9 @@ export function CreateDatabaseForm(props: { databaseId?: string; onForceRerender
     if (!databaseId || !newDatabaseView) return;
 
     setDatabaseView(newDatabaseView);
-    forceRerender();
     showToast({
       title: "View Updated",
     });
-    storeDatabaseView(databaseId, newDatabaseView);
   }
 
   // Get Title Property
@@ -238,7 +146,7 @@ export function CreateDatabaseForm(props: { databaseId?: string; onForceRerender
   });
   return (
     <Form
-      isLoading={isLoading}
+      isLoading={isLoading || isLoadingDatabases}
       navigationTitle={presetDatabaseId ? " →  Create New Page" : "Create Database Page"}
       actions={
         <ActionPanel>
@@ -258,18 +166,23 @@ export function CreateDatabaseForm(props: { databaseId?: string; onForceRerender
             <ActionSetVisibleProperties
               key="set-visible-inputs"
               databaseProperties={databasePropertiesButTitle || []}
-              selectedPropertiesIds={databaseView?.create_properties}
+              selectedPropertiesIds={databaseView?.create_properties || databaseProperties.map((x) => x.id)}
               onSelect={function (propertyId: string) {
-                const databaseViewCopy = databaseView ? JSON.parse(JSON.stringify(databaseView)) : {};
-                databaseViewCopy?.create_properties.push(propertyId);
+                const databaseViewCopy = (databaseView ? JSON.parse(JSON.stringify(databaseView)) : {}) as DatabaseView;
+                if (!databaseViewCopy.create_properties) {
+                  databaseViewCopy.create_properties = databaseProperties.map((x) => x.id);
+                }
+                databaseViewCopy.create_properties.push(propertyId);
                 saveDatabaseView(databaseViewCopy);
               }}
               onUnselect={function (propertyId: string) {
-                const databaseViewCopy = databaseView ? JSON.parse(JSON.stringify(databaseView)) : {};
-                const newVisiblePropertiesIds = databaseViewCopy?.create_properties.filter(function (pid: string) {
-                  return pid !== propertyId;
-                });
-                databaseViewCopy.create_properties = newVisiblePropertiesIds ? newVisiblePropertiesIds : [];
+                const databaseViewCopy = (databaseView ? JSON.parse(JSON.stringify(databaseView)) : {}) as DatabaseView;
+                if (!databaseViewCopy.create_properties) {
+                  databaseViewCopy.create_properties = databaseProperties.map((x) => x.id);
+                }
+                databaseViewCopy.create_properties = databaseViewCopy.create_properties.filter(
+                  (pid) => pid !== propertyId
+                );
                 saveDatabaseView(databaseViewCopy);
               }}
             />
@@ -309,20 +222,20 @@ export function CreateDatabaseForm(props: { databaseId?: string; onForceRerender
       />
       {databaseProperties
         ?.filter(function (dp) {
-          return dp?.id !== "title" && databaseView?.create_properties?.includes(dp?.id);
+          return (
+            dp.id !== "title" &&
+            dp.type !== "title" &&
+            (!databaseView?.create_properties || databaseView.create_properties.includes(dp.id))
+          );
         })
         .sort(function (dpa, dpb) {
-          const value_a = databaseView?.create_properties?.indexOf(dpa?.id);
-          const value_b = databaseView?.create_properties?.indexOf(dpb?.id);
-
-          if (!value_a || value_a === -1) return 1;
-
-          if (!value_b || value_b === -1) return -1;
-
+          if (!databaseView?.create_properties) {
+            return 0;
+          }
+          const value_a = databaseView.create_properties.indexOf(dpa.id);
+          const value_b = databaseView.create_properties.indexOf(dpb.id);
           if (value_a > value_b) return 1;
-
           if (value_a < value_b) return -1;
-
           return 0;
         })
         .map((dp) => {