+++ extensions/notion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Notion Changelog
 
+## [OAuth] - 2022-05-10
+
+- You can now logged in using OAuth instead of having to create your own integration
+- Added an Action to append some content to an existing page
+- The Search command now returns Databases as well. From there, you can navigate to a Database to list all its pages, search in it, have create custom views for each of them.
+
 ## [Improvement to the Create Database Page Command] - 2022-04-29
 
 - Added an Action to open the Notion Page after creating it